wip: add `<Slider>` component
This initial implementation used the wheel event to detect when the user wanted to scroll the slider so we could programmatically show the previous or next items. The problem with this approach is that using the trackpad on OSX will continuously trigger this event due to the scrolling inertia set by the OS. Since detecting whether the event was triggered due to direct user action is almost impossible, we need to come up with a different approach.