style: replace Inter font with system font stack
The previous solution for incluiding custom fonts was relying on CRA's webpack config in order to correctly resolve the font files. This was causing issues with SSR in development since we don't use webpack for our server. I decided to go back to the "system font stack" because it's not that much different from the "Inter" font we were using and it eliminates the need for extra steps.